fix: Correct the test DQL code for Part III (users)
After reading the instructions again, and checking a few questions on
the Knowledge forum, I realised that my username migration was not
complete, as I didn't realise that there were users in the database
that existed only in either (or both) of the "upvotes" and "downvotes"
columns in the "bad_posts" table.

After realising this error, I redid the DQL code that will be used in
the DML to migrate the users once I reset all the data in the workspace.