feat: Add new test DQL for Part III (topics)
After running into issues later in the project, I have decided to keep
things simple and just migrate the topic data without making any format
changes to the topic names themselves.

The reason for this is that, once the posts data is migrated, if the
topic names were altered (even just for format reasons), only 39 369 rows
of data would be migrated, instead of the full 50 000. I also don't
want the alteration of the topic data to be a potential ground on which
my project could be rejected.